release 1.2.11-SNAPSHOT

+++ src/main/java/com/thinkgem/jeesite/modules/cms/web/CommentController.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.thinkgem.jeesite.common.config.Global;
 import com.thinkgem.jeesite.common.persistence.Page;
 import com.thinkgem.jeesite.common.utils.StringUtils;
 import com.thinkgem.jeesite.common.web.BaseController;
@@ -73,6 +74,11 @@ public class CommentController extends BaseController {
 	@RequiresPermissions("cms:comment:edit")
 	@RequestMapping(value = "delete")
 	public String delete(Comment comment, @RequestParam(required=false) Boolean isRe, RedirectAttributes redirectAttributes) {
+
+		if(Global.isDemoMode()){
+			addMessage4DemoMode(redirectAttributes);
+			return "redirect:" + adminPath + "/cms/comment/?repage&delFlag=2";
+		}
 		commentService.delete(comment, isRe);
 		addMessage(redirectAttributes, (isRe!=null&&isRe?"恢复审核":"删除")+"评论成功");
 		return "redirect:" + adminPath + "/cms/comment/?repage&delFlag=2";
